delete: contacts & fix: services-type
