+++ src/components/elements/ServiceCard.tsx
@@ -22,7 +22,7 @@ import {
 } from "react-icons/si";
 import { BiLogoTypescript, BiLogoJavascript } from "react-icons/bi";
 import { BsFiletypeSql } from "react-icons/bs";
-import { getSkillLevelDescription } from "@/types/serviceSkill";
+import { getSkillLevelDescription } from "@/types/service";
 
 type TechStack = {
 	Icon: IconType;
