+++ src/components/elements/ServiceSection.tsx
@@ -4,14 +4,15 @@ import React, { useEffect, useState } from "react";
 import { ServiceCard } from "@/components/elements/ServiceCard";
 import { FaDatabase, FaLaptopCode } from "react-icons/fa";
 import { IoAnalyticsSharp } from "react-icons/io5";
-import { getSkills } from "@/lib/firebase/store/services";
-import { DEFAULT_SERVICES } from "@/constants/services";
+import { RxReload } from "react-icons/rx";
 
-import type { Service } from "@/types/services";
-import type { Skill } from "@/types/skills";
+import { DEFAULT_SERVICES } from "@/constants/services";
+import { getSkills } from "@/lib/firebase/store/services";
+
+import type { Service, Skill } from "@/types/service";
 
 interface Props {
-	initialSkills: Skill[] | null;
+	initialSkills: Skill[];
 }
 
 const ServiceSection: React.FC<Props> = ({ initialSkills }) => {
@@ -53,11 +54,10 @@ const ServiceSection: React.FC<Props> = ({ initialSkills }) => {
 						key={service.id}
 						title={service.title}
 						Icon={
-							service.id === "data-infrastructure"
-								? FaDatabase
-								: service.id === "web-development"
-									? FaLaptopCode
-									: IoAnalyticsSharp
+							service.id === "data-infrastructure" ? FaDatabase :
+							service.id === "web-development" ? FaLaptopCode :
+							service.id === "data-analytics" ? IoAnalyticsSharp :
+							RxReload
 						}
 						items={service.items}
 						skills={serviceSkills}
