--- src/types/skills.ts
@@ -1,6 +1,0 @@
-export type Skill = {
-	id: string;
-	name: string;
-	level: number;
-	category: "データ基盤構築" | "Webサイト開発" | "データ分析・可視化";
-};
